Update 16_swap_chain_recreation.cpp
I wander if the code  of framebufferResizeCallback written by the author is right, because app is a local variable in the function named framebufferResizeCallback, the attributes of which changes never effect the attributes of this instance, and the function does not return anything. I think if change the framebufferResized into static variable is better.